split scenarios from endtoendserver
